General code cleanup.
o removed old stuff no longer used
o fixed some "eclipse" warnings
o fixed a lot of autoboxing/casting stuff
o used FindBugs and actually found 1 real bug 
  (XaLogicalLog leaked in a hashmap for read only tx's during recovery)


git-svn-id: https://svn.neo4j.org/projects/neo/trunk@59 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ community/src/test/unit/neo/store/TestIdGenerator.java
@@ -421,8 +421,8 @@ public class TestIdGenerator extends TestCase
 				float rIndex = random.nextFloat();
 				if ( rIndex < releaseIndex && currentIdCount > 0 )
 				{
-					idGenerator.freeId( ( ( Integer ) idsTaken.remove(  
-						random.nextInt( currentIdCount ) ) ).intValue() ); 
+					idGenerator.freeId( idsTaken.remove(  
+						random.nextInt( currentIdCount ) ).intValue() ); 
 					currentIdCount--;
 				}
 				else
